bajas de metadatos de libros y de admin
Se agregaron las validaciones para eliminar los metadatos del libro (con las comprobaciones de que esos datos no esten en uso) y se puede hacer la baja del admin

+++ app/Http/Controllers/AutorController.php
@@ -112,6 +112,10 @@ class AutorController extends Controller
     public function destroy($id)
     {
         $autor = Autor::findOrFail($id);
+        if ($autor->libros()->count() > 0) {
+            return back()->with('mensaje', 'Este autor esta en uso');
+        }
+
         $autor->delete();
 
         return back()->with('mensaje', 'Autor Eliminado!');
